+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ---
 
 ### New
+* [#58](https://gitlab.com/gitlab-data/permifrost/-/issues/58) Add support for --ignore-membership flag to skip grant/revoke of role memberships
 
 ### Changes